front/syn_metadata: move pure path/intrinsic classifiers + cast_builtin_name out of front/ast (Slice 2)
