+++ majit/majit-translate/src/front/ast.rs
@@ -21,11 +21,14 @@ pub use crate::front::syn_metadata::{
     full_type_string, qualified_full_type_string, qualified_full_type_string_with_imports,
     trait_object_root_name,
 };
+pub(crate) use crate::front::syn_metadata::is_synthetic_unit_variant_path;
 use crate::front::syn_metadata::{
-    array_item_value_type_from_array_type_id, bare_type_root_from_type_str,
-    dyn_trait_root_from_type_str, extract_element_type_from_str, method_as_ref_return_type,
-    outer_generic_inner_type, split_tuple_type_elements, transparent_option_inner_type,
-    transparent_result_err_type, type_root_from_type_string, type_string_to_value_type,
+    array_item_value_type_from_array_type_id, bare_type_root_from_type_str, cast_builtin_name,
+    dyn_trait_root_from_type_str, extract_element_type_from_str, intrinsic_call_result_type,
+    is_synthetic_ctor_path, is_synthetic_result_option_wrapper_path, kind_char_to_value_type,
+    method_as_ref_return_type, outer_generic_inner_type, path_as_value_float_constant,
+    split_tuple_type_elements, transparent_option_inner_type, transparent_result_err_type,
+    type_root_from_type_string, type_string_to_value_type,
 };
 
 /// Result of lowering one expression or statement-list tail.
@@ -7118,94 +7121,6 @@ fn canonical_call_target(expr: &syn::Expr, ctx: &GraphBuildContext) -> CallTarge
     }
 }
 
-/// Decide whether `segments` should be lowered as
-/// `CallTarget::SyntheticTransparentCtor` instead of `FunctionPath`.
-///
-/// Two disjoint groups share the same routing decision but receive
-/// different downstream treatment:
-///
-/// * **Result/Option elision whitelist** (`Ok`/`Err`/`Some`):
-///   - flowspace_adapter routes via `HostObject::new_class(name, [])`
-///     → `ClassDesc` → `SomeInstance(classdef)`.
-///   - jtransform `is_synthetic_result_option_ctor` then re-elides the
-///     single-arg call back to its inner value, matching upstream
-///     `lloperation:cast_to_ptr` semantics for `Ok(x)` / `Err(e)` /
-///     `Some(v)`.
-///
-/// * **Pyre-side `Class::Variant` ctors** (`LoopResult::Done` etc.):
-///   - flowspace_adapter routes identically via
-///     `HostObject::new_class(name, [])` → `SomeInstance(classdef)`.
-///   - jtransform does **not** elide them (they are real values, not
-///     transparent wrappers); the call lowers to a normal instance
-///     construction.
-///   - Upstream RPython gets the same shape by lifting each enum
-///     variant as a Class with a registered `__init__` FunctionDesc;
-///     pyre's surface DSL lacks that lifting pass today so the route
-///     is gated explicitly until a `sym_enum_variant` analyser pass
-///     lands.
-///
-/// Both groups must skip the `PyreCallRegistry` function-path lookup
-/// because neither is registered there; the discriminator is the
-/// caller's check at `canonical_call_target` that
-/// `registered_function_path` returns false.
-fn is_synthetic_ctor_path(segments: &[String]) -> bool {
-    is_synthetic_result_option_wrapper_path(segments) || is_synthetic_unit_variant_path(segments)
-}
-
-/// `Ok`/`Err`/`Some` and their qualified spellings.  These are
-/// always one-argument transparent wrappers that
-/// `jtransform::is_synthetic_result_option_ctor` elides at the
-/// `args.len() == 1` site.  Valid only as call targets — using the
-/// bare ctor name as a value would lower to a 0-arg synthetic call
-/// the elision pass does not handle.
-fn is_synthetic_result_option_wrapper_path(segments: &[String]) -> bool {
-    let path: Vec<&str> = segments.iter().map(String::as_str).collect();
-    matches!(
-        path.as_slice(),
-        ["Ok"]
-            | ["Err"]
-            | ["Some"]
-            | ["Result", "Ok"]
-            | ["Result", "Err"]
-            | ["Option", "Some"]
-            | ["result", "Result", "Ok"]
-            | ["result", "Result", "Err"]
-            | ["option", "Option", "Some"]
-            | ["std", "result", "Result", "Ok"]
-            | ["std", "result", "Result", "Err"]
-            | ["std", "option", "Option", "Some"]
-            | ["core", "result", "Result", "Ok"]
-            | ["core", "result", "Result", "Err"]
-            | ["core", "option", "Option", "Some"]
-    )
-}
-
-/// Path-as-value numeric constants — Rust counterpart of PyPy
-/// `flowspace.LOAD_GLOBAL` resolving a statically-known module
-/// attribute to a `Constant(value)` node.  Returns the `f64` bit
-/// pattern matching the `syn::Lit::Float` arm's
-/// `OpKind::ConstFloat(v.to_bits())` form.
-///
-/// Only `f64::{INFINITY, NEG_INFINITY, NAN}` and their long forms
-/// `{std,core}::f64::*` are recognised here; other numeric path
-/// constants would each need their own typed `Const*` op (the
-/// `OpKind::ConstFloat` variant is float-specific).
-fn path_as_value_float_constant(segments: &[String]) -> Option<u64> {
-    let path: Vec<&str> = segments.iter().map(String::as_str).collect();
-    let leaf = match path.as_slice() {
-        ["f64", leaf] => leaf,
-        ["std", "f64", leaf] => leaf,
-        ["core", "f64", leaf] => leaf,
-        _ => return None,
-    };
-    match *leaf {
-        "INFINITY" => Some(f64::INFINITY.to_bits()),
-        "NEG_INFINITY" => Some(f64::NEG_INFINITY.to_bits()),
-        "NAN" => Some(f64::NAN.to_bits()),
-        _ => None,
-    }
-}
-
 /// Resolve a path expression against the program-wide
 /// `pub const` / `pub static` table on `ctx.module_statics` and
 /// return its compile-time literal value when known.  Mirrors how
@@ -7300,32 +7215,6 @@ fn lookup_module_static_literal(
         }
     }
     None
-}
-
-/// Pyre-side `Class::Variant` unit-variant ctors.  These are valid
-/// as bare path-expression values; `flowspace_adapter` pre-folds them
-/// to `Hlvalue::Constant(ConstValue::HostObject(prebuilt_instance))`
-/// before the rtyper sees a call (mirrors PyPy `rtyper` resolving
-/// `SomePBC([InstanceDesc(<unit-variant>)])` to a singleton constant
-/// before `jtransform`).  Exposed `pub(crate)` so
-/// `translator::rtyper::flowspace_adapter::is_synthetic_unit_variant_call`
-/// reads the same allowlist.
-pub(crate) fn is_synthetic_unit_variant_path(segments: &[String]) -> bool {
-    let path: Vec<&str> = segments.iter().map(String::as_str).collect();
-    matches!(
-        path.as_slice(),
-        ["LoopResult", "Done"]
-            | ["LoopResult", "ContinueRunningNormally"]
-            | ["JitAction", "Return"]
-            | ["JitAction", "Continue"]
-            | ["StepResult", "Continue"]
-            | ["CompareOp", "Lt"]
-            | ["CompareOp", "Le"]
-            | ["CompareOp", "Gt"]
-            | ["CompareOp", "Ge"]
-            | ["CompareOp", "Eq"]
-            | ["CompareOp", "Ne"]
-    )
 }
 
 fn registered_function_path(segments: &[String], ctx: &GraphBuildContext) -> bool {
@@ -7672,272 +7561,6 @@ fn primitive_method_result_type(
         _ => None,
     }
 }
-
-fn intrinsic_call_result_type(segments: &[String]) -> Option<ValueType> {
-    let path: Vec<&str> = segments.iter().map(String::as_str).collect();
-    match path.as_slice() {
-        // Rust's `size_of::<T>()` is a compile-time `usize`.  RPython
-        // carries such layout sizes as `lltype.Signed` constants before
-        // codewriter, so the equivalent register class is Int.
-        ["std", "mem", "size_of"] | ["core", "mem", "size_of"] | ["mem", "size_of"] => {
-            Some(ValueType::Int)
-        }
-        // Associated f64 helpers used by the floatobject.py port.
-        ["f64", "copysign"] | ["std", "f64", "copysign"] | ["core", "f64", "copysign"] => {
-            Some(ValueType::Float)
-        }
-        _ => None,
-    }
-}
-
-fn kind_char_to_value_type(kind: char) -> Option<ValueType> {
-    match kind {
-        'i' => Some(ValueType::Int),
-        'r' => Some(ValueType::Ref(None)),
-        'f' => Some(ValueType::Float),
-        'v' => Some(ValueType::Void),
-        _ => None,
-    }
-}
-
-/// RPython equivalent of Rust's `expr as T` for numeric / Bool /
-/// pointer casts.
-///
-/// Upstream has no `as T` syntax — numeric conversions are spelled as
-/// builtin calls `int(v)` / `float(v)` / `bool(v)` (`rbuiltin.py:178-189`),
-/// each of which the rtyper routes through
-/// `BuiltinFunctionRepr.rtype_simple_call → rtype_builtin_int/float/bool
-/// → arg_repr.rtype_int/float/bool` per `Repr` trait method.  Those
-/// per-Repr methods (rint.py:137-147, rfloat.py:48-58, rbool.py:55-70)
-/// emit the low-level `cast_*_to_*` / `*_is_true` op as a side effect
-/// of `inputargs(target_lltype)` coercion through the appropriate
-/// `pair_X_Y_convert_from_to` (rint.py:645-675, rbool.py:49-84).
-///
-/// Pointer↔int conversions go through `lltype.cast_ptr_to_int(p)` /
-/// `lltype.cast_int_to_ptr(T, v_int)` (`rbuiltin.py:543-557`), routed
-/// through the same `BUILTIN_TYPER` registry via the module-qualified
-/// HOST_ENV resolver in `flowspace_adapter`.
-///
-/// Maps each `(source, target)` pair to the matching callable path —
-/// a single `__builtin__` name (one segment) for numeric/Bool casts,
-/// or a fully qualified `["rpython", "rtyper", "lltypesystem",
-/// "lltype", <attr>]` / `["rpython", "rlib", "rarithmetic", <attr>]`
-/// path for pointer / Unsigned-bridging casts.  Pairs not covered
-/// (identity casts, source-type-unknown casts) return `None`, letting
-/// the caller emit `OpKind::UnaryOp { op: "same_as", result_ty, .. }`
-/// to preserve target-type propagation through the graph
-/// (rtyper.py:478-481 internal renaming op).  All 13 typed numeric /
-/// ptr / Unsigned cast names retired across Slices A.3 / B.1 / A.4a /
-/// A.4b / A.4c.
-fn cast_builtin_name(
-    source_ty: Option<&ValueType>,
-    target_ty: &ValueType,
-) -> Option<&'static [&'static str]> {
-    match (source_ty, target_ty) {
-        // upstream `float(v)`:
-        //   `IntegerRepr.rtype_float` (rint.py:144-147) — `inputargs(Float)`
-        //   coerces via `pair(IntegerRepr, FloatRepr).convert_from_to`
-        //   (rint.py:645-655) → `genop('cast_int_to_float', ...)`.
-        //   `BoolRepr.rtype_float` (rbool.py:65-70) — `inputargs(Float)`
-        //   coerces via `pair(BoolRepr, FloatRepr).convert_from_to`
-        //   (rbool.py:49-56) → `genop('cast_bool_to_float', ...)`.
-        (Some(ValueType::Int), ValueType::Float) => Some(&["float"]),
-        (Some(ValueType::Bool), ValueType::Float) => Some(&["float"]),
-        // `IntegerRepr.rtype_float` (rint.py:144-147) with `self`
-        // having lltype `Unsigned` — `inputargs(Float)` coerces via
-        // `pair(IntegerRepr, FloatRepr).convert_from_to`
-        // (rint.py:645-655) and the `r_from.lowleveltype == Unsigned`
-        // arm emits `genop('cast_uint_to_float', ...)`.
-        (Some(ValueType::Unsigned), ValueType::Float) => Some(&["float"]),
-        // upstream `int(v)`:
-        //   `FloatRepr.rtype_int` (rfloat.py:48-53) —
-        //   `genop('cast_float_to_int', ...)`.
-        //   `BoolRepr.rtype_int` (rbool.py:55-60) — `inputargs(Signed)`
-        //   coerces via `pair(BoolRepr, IntegerRepr).convert_from_to`
-        //   (rbool.py:73-78) → `genop('cast_bool_to_int', ...)`.
-        (Some(ValueType::Float), ValueType::Int) => Some(&["int"]),
-        (Some(ValueType::Bool), ValueType::Int) => Some(&["int"]),
-        // upstream `intmask(r_uint(...))` (rbuiltin.py:220-225) —
-        // `IntegerRepr.rtype_int` (rint.py:137-142) forbids `int(v)`
-        // for `r_uint`, so the canonical Unsigned→Signed entry is
-        // `rarithmetic.intmask(v_uint)`.  `rtype_intmask` calls
-        // `hop.inputargs(Signed)` which coerces from Unsigned via
-        // `pair(IntegerRepr, IntegerRepr).convert_from_to`
-        // (rint.py:202-213) emitting `cast_uint_to_int`.
-        (Some(ValueType::Unsigned), ValueType::Int) => {
-            Some(&["rpython", "rlib", "rarithmetic", "intmask"])
-        }
-        // upstream `bool(v)`:
-        //   `IntegerRepr.rtype_bool` (rint.py:85-88) —
-        //   `genop(self.opprefix + 'is_true', ...)`.
-        //   `FloatRepr.rtype_bool` (rfloat.py:32-34) —
-        //   `genop('float_is_true', ...)`.
-        (Some(ValueType::Int), ValueType::Bool) => Some(&["bool"]),
-        (Some(ValueType::Float), ValueType::Bool) => Some(&["bool"]),
-        // `IntegerRepr.rtype_bool` (rint.py:85-88) — `genop(self.\
-        // opprefix + 'is_true', ...)` — emits `uint_is_true` for
-        // Unsigned-prefixed Repr.
-        (Some(ValueType::Unsigned), ValueType::Bool) => Some(&["bool"]),
-        // upstream `lltype.cast_ptr_to_int(p)` (rbuiltin.py:543-548) —
-        //   `genop('cast_ptr_to_int', vlist, resulttype=Signed)`.
-        // upstream `lltype.cast_int_to_ptr(T, v_int)`
-        // (rbuiltin.py:551-557) — `genop('cast_int_to_ptr', [v_input],
-        //   resulttype=hop.r_result.lowleveltype)`.
-        //
-        // TODO: the upstream surface is
-        // a 2-arg call `simple_call(lltype.cast_int_to_ptr, PTRTYPE,
-        // oddint)` where `PTRTYPE` is a constant Ptr type marker
-        // (`ann_cast_int_to_ptr` asserts `PtrT.is_constant()` and
-        // returns `SomePtr(ll_ptrtype=PtrT.const)`,
-        // lltype.py:2379-2382).  Pyre emits a 1-arg call here.
-        //
-        // The constant carrier exists — `ConstValue::LowLevelType(Box<\
-        // LowLevelType>)` (flowspace/model.rs:1958) wraps any lltype
-        // including `LowLevelType::Ptr(Box<Ptr>)` — but the **frontend**
-        // does not yet have the concrete Ptr lltype at cast-lowering
-        // time.  `Expr::Cast` carries `ValueType::Ref` (the opaque
-        // high-level surface type); the concrete Ptr is computed only
-        // later by the rtyper from the result variable's annotation.
-        // The result lltype is recovered from `hop.r_result.\
-        // lowleveltype` at rtype-time — the same Ptr upstream would
-        // have read from `PtrT.const`.  Graduating to the 2-arg form
-        // is blocked on threading the concrete result Ptr from the
-        // rtyper-side annotation back to the frontend (or a deferred
-        // 2-arg rewrite in jtransform / annotator after the Ptr is
-        // known).
-        (Some(ValueType::Ref(_)), ValueType::Int) => Some(&[
-            "rpython",
-            "rtyper",
-            "lltypesystem",
-            "lltype",
-            "cast_ptr_to_int",
-        ]),
-        (Some(ValueType::Int), ValueType::Ref(_)) => Some(&[
-            "rpython",
-            "rtyper",
-            "lltypesystem",
-            "lltype",
-            "cast_int_to_ptr",
-        ]),
-        // upstream `r_uint(v)` (rarithmetic.py:600
-        // `build_int('r_uint', False, LONG_BIT)`) → the
-        // `ForTypeEntry.specialize_call` body emits
-        // `inputargs(Unsigned)` which coerces the source via
-        // `pair(SrcRepr, IntegerRepr<Unsigned>).convert_from_to`
-        // (rint.py:202-213 / rint.py:657-675 / rbool.py:62-66)
-        // emitting `cast_int_to_uint` / `cast_float_to_uint` /
-        // `cast_bool_to_uint`.
-        (Some(ValueType::Int), ValueType::Unsigned)
-        | (Some(ValueType::Float), ValueType::Unsigned)
-        | (Some(ValueType::Bool), ValueType::Unsigned) => {
-            Some(&["rpython", "rlib", "rarithmetic", "r_uint"])
-        }
-        _ => None,
-    }
-}
-
-/// Maps a primitive numeric type name to its `ValueType` for the
-/// `<prim>::from` coercion path.  Mirrors the primitive arms of
-/// `classify_fn_arg_ty` (u* → `Unsigned` so `cast_builtin_name` selects
-/// the `r_uint` / `intmask` bridges, not the int identity).  Returns
-/// `None` for non-primitive receivers so `Foo::from` stays an ordinary
-/// call.
-fn numeric_type_name_to_value_type(name: &str) -> Option<ValueType> {
-    match name {
-        "i8" | "i16" | "i32" | "i64" | "isize" => Some(ValueType::Int),
-        "u8" | "u16" | "u32" | "u64" | "usize" => Some(ValueType::Unsigned),
-        "f32" | "f64" => Some(ValueType::Float),
-        _ => None,
-    }
-}
-
-/// Recognizes `<prim>::from` — the infallible numeric widening
-/// conversion.  Returns the target `ValueType` when `path` is exactly
-/// `<primitive numeric type>::from`, else `None`.  `try_from` is
-/// excluded: its `Result<T, E>` result needs the Ok-elision / unwrap
-/// composition handled elsewhere.
-fn numeric_from_target_type(path: &syn::Path) -> Option<ValueType> {
-    let segments: Vec<String> = path.segments.iter().map(|s| s.ident.to_string()).collect();
-    match segments.as_slice() {
-        [ty, method] if method.as_str() == "from" => numeric_type_name_to_value_type(ty),
-        _ => None,
-    }
-}
-
-/// Recognizes `std::ptr::eq` / `core::ptr::eq` / `ptr::eq` — pointer
-/// identity comparison.  RPython spells this as `a is b`, lowered to
-/// `ptr_eq` (rmodel.py:300 rtype_is_); pyre produces the same op from a
-/// `BinOp { op: "eq" }` on two Ref operands, which jtransform rewrites
-/// Ref==Ref → `ptr_eq` (jtransform.rs:849 / jtransform.py:1243
-/// rewrite_op_ptr_eq).
-fn is_ptr_eq_path(path: &syn::Path) -> bool {
-    let segments: Vec<String> = path.segments.iter().map(|s| s.ident.to_string()).collect();
-    match segments.as_slice() {
-        [a, b, c] => {
-            matches!(a.as_str(), "std" | "core") && b.as_str() == "ptr" && c.as_str() == "eq"
-        }
-        [a, b] => a.as_str() == "ptr" && b.as_str() == "eq",
-        _ => false,
-    }
-}
-
-/// Lowers a numeric / Bool / pointer value cast of `operand_var` (whose
-/// source annotation is `source_ty`) to `result_ty`.  Routes through the
-/// canonical `int()` / `float()` / `bool()` / `cast_*` builtin-call
-/// chain (`cast_builtin_name`, rbuiltin.py:178-189) when a typed
-/// coercion is required, else emits the identity-preserving `same_as`
-/// (rtyper.py:478-481).  Shared by `Expr::Cast` (`x as T`) and the
-/// `<prim>::from(x)` function-call spelling, which RPython expresses
-/// with the same `int(v)` / `r_uint(v)` builtin calls.
-fn lower_value_cast(
-    graph: &mut FunctionGraph,
-    block: BlockId,
-    operand_var: crate::flowspace::model::Variable,
-    source_ty: Option<ValueType>,
-    result_ty: ValueType,
-) -> crate::flowspace::model::Variable {
-    if let Some(segments) = cast_builtin_name(source_ty.as_ref(), &result_ty) {
-        return graph
-            .push_op_var(
-                block,
-                OpKind::Call {
-                    target: CallTarget::FunctionPath {
-                        segments: segments.iter().map(|s| s.to_string()).collect(),
-                    },
-                    args: vec![operand_var],
-                    result_ty,
-                },
-                true,
-            )
-            .expect("OpKind::Call has has_result=true");
-    }
-    graph
-        .push_op_var(
-            block,
-            OpKind::UnaryOp {
-                op: "same_as".to_string(),
-                operand: operand_var,
-                result_ty,
-            },
-            true,
-        )
-        .expect("OpKind::UnaryOp has has_result=true")
-}
-
-// `cast_op_name` retired in Slice C.1 — after the numeric / Bool /
-// pointer / Unsigned cast families all routed through `cast_builtin_\
-// name → simple_call(<host_callable>, v)` (Slices A.3 / B.1 /
-// A.4a-c), the only remaining arm was the `_ => "same_as"` catch-all
-// for identity / source-type-unknown casts.  Slice F3 restored the
-// `same_as` fallback inside `Expr::Cast` directly: when
-// `cast_builtin_name` returns None, the lowering emits
-// `OpKind::UnaryOp { op: "same_as", result_ty, .. }` so the target
-// `ValueType` propagates through the graph (`graph_value_type` reads
-// the producer op's result_ty).  Downstream
-// `rbuiltin::rtype_same_as` (verbatim port of `rtyper.py:478-481`)
-// preserves the operand's lltype; `backendopt::removenoops::remove_\
-// same_as` (removenoops.py:47-48) collapses the identity copy when
-// safe.
 
 fn lookup_method_return_type<'a>(
     ctx: &'a GraphBuildContext,
